Add files via upload
A WAY faster solution. (check comparison_between_solutions.txt for details)

I changed the way we save the data and the way we access it. 
Now using one large dictionary(), and accessing the data using the street's name instead of always finding the index in the list (which took a time).

Another very important change: I changed the way we count cars passing in the street.
Instead of iterating for each street on all the cars and checking the path of each car, we iterate once on all the cars and up the count for each street we see on the path.
We have a counter for each street for passing cars in the streets dictionary().